pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arxiv-recommendation"
version = "0.1.0"
description = "Personal arXiv recommendation system with LLM embeddings"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    # Core OpenAI integration
    "openai>=1.52.0",
    
    # Data handling
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "h5py>=3.10.0",
    "filelock>=3.13.0",
    
    # Multi-agent framework (lightweight)
    "autogen-agentchat>=0.4.0",
    
    # Web interface
    "streamlit>=1.39.0",
    
    # Optional performance (for larger datasets)
    "faiss-cpu>=1.8.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",  # Beautiful CLI output
    "typer>=0.12.0", # CLI framework
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    
    # Code quality
    "ruff>=0.6.0",      # Fast linter & formatter
    "mypy>=1.8.0",      # Type checking
    
    # Development tools
    "pre-commit>=3.6.0",
    "jupyter>=1.0.0",   # For data exploration
]

web = [
    "streamlit>=1.39.0",
    "plotly>=5.17.0",   # Interactive charts
    "altair>=5.2.0",    # Streamlit-native charts
]

performance = [
    "faiss-cpu>=1.7.0", # CPU-based similarity search  
    "redis>=5.1.0",     # Advanced caching
]

[project.scripts]
arxiv-rec = "main:main"
arxiv-sync = "scripts.daily_sync:main"  
arxiv-setup = "scripts.setup_database:main"
arxiv-migrate-embeddings = "scripts.migrate_embeddings:main"
arxiv-cache-stats = "scripts.check_cache_stats:main"
arxiv-web = "run_web:main"

[project.urls]
Homepage = "https://github.com/yourusername/arxiv-recommendation"
Repository = "https://github.com/yourusername/arxiv-recommendation"
Issues = "https://github.com/yourusername/arxiv-recommendation/issues"

# Tool configurations
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]